Use floats instead of 8 bit PCM in Recorder

+++ src/wavwriter.hpp
@@ -4,6 +4,6 @@
 #define _WAV_SYNTH_WAVWRITER_H_INCLUDED
 
 #include <stdint.h>
-void writewav(uint8_t *data, int samples, int sample_rate, const char* filename);
+void writewav(float *data, int samples, int sample_rate, const char* filename);
 
 #endif
